backend/rag_system/requirements.txt
google-genai
faiss-cpu
pdfplumber
python-docx
nltk
sentence-transformers
numpy
python-dotenv
openpyxl
xlrd
